+++ docs/changes.rst
@@ -37,6 +37,8 @@ Released: not yet
 
 * Examples: Fixed errors in and improved metrics examples.
 
+* Fixed issues raised by new Pylint version 2.9.1.
+
 **Enhancements:**
 
 * Added support for 'Console.list_permitted_partitions()' and